Extracting different error types into separate files, WIP forbidden error class
Jira ticket: CAMS-142
