+++ backend/functions/lib/common-errors/cams-error.ts
@@ -1,0 +1,11 @@
+export class CamsError extends Error {
+  status: number;
+  module: string;
+
+  constructor(status: number, message: string, module: string) {
+    super();
+    super.message = message;
+    this.status = status;
+    this.module = module;
+  }
+}
