+++ backend/functions/lib/use-cases/assignment.exception.ts
@@ -1,7 +1,7 @@
-import { CamsError } from '../cams-error';
+import { CamsError } from '../common-errors/cams-error';
 
 export class AssignmentException extends CamsError {
-  constructor(status: number, message: string) {
-    super(status, message, 'ASSIGNMENTS');
+  constructor(status: number, message: string, module: string = 'ASSIGNMENTS') {
+    super(status, message, module);
   }
 }
